% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAlignments}
\alias{addAlignments}
\title{Add multiple sequence alignment to guideSet object}
\usage{
addAlignments(guideSet, files = NULL, max_gap_freq = 0.8,
  force = FALSE)
}
\arguments{
\item{guideSet}{guideSet with targets}

\item{file}{String}
}
\description{
Either calculates multiple sequence alignment and consensus sequence or imports from external file
}
