% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addGuides}
\alias{addGuides}
\title{Add guides to a guideSet object.}
\usage{
addGuides(guideSet, n_mismatches = 0, guide_length = 19,
  gc_content = c(0.4, 0.8), min_Son = 10, max_Soff = Inf,
  consensus_range = NULL, n_clust = 15, method = "max", coeff = 1,
  PAM = "NGG", force = FALSE)
}
\arguments{
\item{guideSet}{guideSet object with targets.}

\item{n_mismatches}{Single integer of either 0, 1, 2, or 3. Maximal number of tolerated mismatches when assessing guideRNA binding targets. Defaults to 0.}

\item{guide_length}{Single integer between 10 and 30. Basepair size of the guideRNAs. 
A spacer length of 19 (the default) has been optimal in 'Sequence determinants of improved CRISPR sgRNA design', 
  but shorter guides may result in fewer actual off-targets 'Improving CRISPR-Cas nuclease specificity using truncated guide RNAs'}

\item{consensus_range}{DataFrame with repname, start, and end columns. Discard guideRNAs targeting parts outside of \code{consensus_range} on the consensus.}

\item{PAM}{Character. Currently only 'NGG' PAMs are supported.}

\item{min_targets}{Single integer >= 0. Discard guideRNAs with less than \code{min_targets} hits on target.}

\item{max_off_targets}{Single integer >= 0. Discard guideRNAs with more than \code{max_off_targets} off target hits.}
}
\value{
Returns a guideSet object
}
\description{
Adds all potential guideRNAs against the targets of interest.
}
\examples{
gs <- createGuideSet(Hsapiens)
gs <- addTargets(targets = c('LTR12C', 'LTR12E'), alignment = TRUE)
}
