% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTargets}
\alias{addTargets}
\title{Add targets to a guideSet object.}
\usage{
addTargets(guideSet, targets = NULL, max_gap_freq = 0.8,
  blacklist = FALSE, min_dist = 0, force = FALSE)
}
\arguments{
\item{guideSet}{guideSet object to add targets to.}

\item{targets}{Character or GRanges. Either names of repeat families or GRanges object with target coordinates.}

\item{blacklist}{Boolean. Should targets overlapping cis regulatory regions be blacklisted. False by default.}

\item{alignment}{Boolean. Should a rough multiple sequence alignment and consensus sequence be added? 
Useful for inspecting or targeting specific parts of a family.}

\item{blacklist_dist}{Integer >= 0. Minimum basepair distance of target loci to cis regulatory regions.}
}
\value{
Returns a guideSet object
}
\description{
Adds target loci you want to design guideRNAs against.
}
\examples{
gs <- createGuideSet(Hsapiens)
gs <- addTargets(targets = c('LTR12C', 'LTR12E'), alignment = TRUE)
}
