% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guideSet.R
\name{createGuideSet}
\alias{createGuideSet}
\title{Creates a new guideSet object}
\usage{
createGuideSet(genome, alt_chromosomes = FALSE, tes = GRanges(),
  cis = GRanges(), n_cores = NULL, refdir = "", seed = 19)
}
\description{
Creates a new guideSet object
}
