% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{essential_genes}
\alias{essential_genes}
\title{Essential human gene promoter coordinates}
\format{
bed file with 13944 genomic intervals (each 2 kb in size).
}
\source{
\href{https://www.gencodegenes.org/human/}{Gencode v28 annotation}, and essential genes \href{http://tko.ccbr.utoronto.ca/Data/core-essential-genes-sym_HGNCID}{Hart et al., 2015}
}
\description{
Genomic coordinates (hg38) of essential human gene promoters. Transcription start sites of essential genes (Hart et al., 2015) were determined from Gencode annotation (v28) and extended 1 kb up- and downstream.
}
\section{hg38_essential_gene_tss_coords.bed}{


This data is used to exemplify \code{blacklist} of the \code{createGuideSet} function.
}

