% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCombinations}
\alias{plotCombinations}
\title{Generates combinations QC plots}
\usage{
plotCombinations(guideSet)
}
\arguments{
\item{guideSet}{guideSet containing combinations.}
}
\value{
list of ggplot objects.
}
\description{
Computes multiple QC plots and arranges into a figure with subpanels:
\describe{
\item{a}{ Table showing total on/off-target scores (Son_tot and Soff_tot) for the best combination per number of guides (columns). }
\item{b}{ Total on- and off-target score per number of greedy search iterations. Colors indicate the number of guides (from yellow (at least 2) to black (defined by \code{max_guides} in the \code{addCombinations} function). }
\item{c}{ Total on- and off-target score per number of guides. }
\item{d}{ Total on- and off-target score for all computated combinations per number of guides. }
\item{e}{ Off-target coverage for the best guide combination per number of guides and off-target type (e.g. transposon family). Colors indicate predicted binding affinity to the off-target. }
\item{f}{ Distance to the nearest cis regulatory feature (defined by \code{cis} in the \code{createGuideSet} function) for the best combination number of guides. }
\item{g}{ On-target coverage for the best guide combination per number of guides and on-target transposon family. Colors indicate predicted binding affinity to the target. }
\item{h}{ Coverage for the best guide combination on the consensus model of transposon families. Colors indicate the number of guides (from yellow (1) to black (\code{max_guides})) }
}
}
\examples{
\dontrun{
gs <- createGuideSet(Hsapiens, tes = te_annotation_df)
gs <- addTargets(gs, targets = 'LTR13A')
gs <- plotTargets(gs)
}
}
\seealso{
\code{\link[=addCombinations]{addCombinations()}}, and \code{\link[=export]{export()}}
}
