% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{repnames}
\alias{repnames}
\title{Returns transposable element families from guideSet}
\usage{
repnames(guideSet, pattern = NULL, fixed = FALSE)
}
\arguments{
\item{guideSet}{guideSet object to query.}

\item{pattern}{Regular expression. Returns only families that match \code{pattern}.}

\item{fixed}{Logical. If \code{TRUE}, \code{pattern} is matched as is.}
}
\value{
Character vector.
}
\description{
Retrieves all available families in the provided guideSet that match defined characteristics.
}
\examples{
\dontrun{
gs <- createGuideSet(genome = Hsapiens, tes = te_annotation_df)
families <- repnames(gs, pattern = 'LTR12') # returns all families containing 'LTR12' in their name.

# Using a pattern not found in guideSet will suggest similar hits, e.g.:
repnames(gs, pattern = 'LTR5HS')
}
}
\seealso{
\code{\link[=createGuideSet]{createGuideSet()}}
}
