% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{selGuides}
\alias{selGuides}
\title{Blacklist guideRNAs}
\usage{
selGuides(
  guideSet,
  min_Son = NULL,
  max_Soff = NULL,
  consensus_range = NULL,
  gc_content = c(0.4, 0.8)
)
}
\arguments{
\item{min_Son}{Numeric. Guides below \code{min_Son} are blacklisted. If \code{NULL} (the default), \code{min_Son} is calculated based on the target score distribution (75th percentile).}

\item{max_Soff}{Numeric. Guides exceeding \code{min_Son} are blacklisted. If \code{NULL} (the default), \code{max_Soff} is calculated based on the off-target score distribution (99.9th percentile).}

\item{consensus_range}{Data.frame with repname, start, and end columns. Scores guideRNA target binding sites outside of provided consensus range neutrally.}

\item{gc_content}{Numeric vector of length two. Elements must be from 0 through 1. For example, c(0.4, 0.8) blacklists guides with GC content not between 40 and 80 percent}
}
\description{
Adds a blacklist flag to guides not fulfilling defined criteria
}
\examples{
\dontrun{
gs <- createGuideSet(Hsapiens, tes = te_annotation_df)
gs <- addTargets(gs, targets = 'LTR13')
gs <- addGuides(gs, guide_length = 16, n_mismatches = 0, gc_content = c(0.25, 0.9), n_clust = 12)
gs <- selGuides(gs, min_Son = 100, gc_content = c(0.1, 0.3))
gs <- plotGuides(gs)
}
}
