% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAlignments}
\alias{addAlignments}
\title{Add multiple sequence alignment to guideSet object}
\usage{
addAlignments(guideSet, files = NULL, max_gap_freq = 0.8,
  iterations = 2, refinements = 1, force = FALSE)
}
\arguments{
\item{guideSet}{guideSet containing targets}

\item{files}{data.frame. Requires columns 'repname' and 'path' storing family identifiers and paths to the multiple sequence alignment, respectively.}

\item{max_gap_freq}{Numeric. Removes positions on alignment with higher gap (-) frequency than \code{max_gap_freq}.}

\item{iterations}{Numeric. Passed to AlignSeqs from DECIPHER package.}

\item{refinements}{Numeric. Passed to AlignSeqs from DECIPHER package.}

\item{force}{Logical. If TRUE, overwrites existing alignments.}
}
\value{
Returns a guideSet object with alignments.
}
\description{
Calculates rough multiple sequence alignment and consensus sequence or imports from external file
}
