% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addCombinations}
\alias{addCombinations}
\title{Add guideRNA combinations to guideSet object.}
\usage{
addCombinations(guideSet, max_guides = 5, greedy = TRUE,
  iterations = 10, method = c("max", "min", "ratio", "regression"),
  coeff = 1, force = FALSE)
}
\arguments{
\item{guideSet}{guideSet object containing guides.}

\item{max_guides}{Numeric. Maximum number of distinct guides to consider when calculating combinations.}

\item{greedy}{Logical. If \code{TRUE} (the default), performs an additional greedy optimization for selected guide combinations.}

\item{iterations}{Integer > 0. Number of greedy search iterations. Usually 10 iterations are sufficient for convergence.}

\item{method}{String. Method of how to pick the best guideRNA per cluster.}

\item{coeff}{Integer. If \code{method} is \code{regression}, \code{coeff} modifies the slope of the linear regression.}

\item{force}{Logical. If \code{TRUE}, overwrites existing results}
}
\value{
Returns a guideSet object with combinations.
}
\description{
Selects set of non-redundant guideRNAs and computes all potential combinations against targets of interest.
}
