% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addCombinations}
\alias{addCombinations}
\title{Add guideRNA combinations to guideSet object.}
\usage{
addCombinations(guideSet, max_guides = 5, greedy = TRUE,
  iterations = 10, method = c("max", "min", "ratio", "regression"),
  coeff = 1, force = FALSE)
}
\arguments{
\item{max_guides}{Numeric. Maximum number of distinct guides to consider when calculating combinations. Do not use a higher number than experimentally feasible.}

\item{greedy}{Logical. If \code{TRUE} (the default), performs an additional greedy search optimization for selected guide combinations.}

\item{iterations}{Integer > 0. Number of greedy search iterations. Usually 10 iterations are sufficient for convergence.}

\item{method}{String. Method of how to pick the best guideRNA per cluster.}

\item{coeff}{Integer. If \code{method} is \code{regression}, \code{coeff} modifies the slope of the linear regression.}

\item{force}{Logical. If \code{TRUE}, overwrites existing combinations.}

\item{guideSet.}{guideSet object containing guides.}
}
\value{
Returns a guideSet object contaning combinations.
}
\description{
Finds the optimal combination of non-redundant guideRNAs.
}
\examples{
\dontrun{
gs <- createGuideSet(Hsapiens, tes = te_annotation_df)
gs <- addTargets(gs, targets = 'LTR13A')
gs <- addGuides(gs, guide_length = 16)
gs <- addCombinations(gs, max_guides = 8)
gs <- plotCombinations(gs)
}
}
\seealso{
\code{\link[=plotCombinations]{plotCombinations()}}, \code{\link[=createGuideSet]{createGuideSet()}}, \code{\link[=addTargets]{addTargets()}}, and \code{\link[=addGuides]{addGuides()}}
}
