% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addGuides}
\alias{addGuides}
\title{Add guides to a guideSet object.}
\usage{
addGuides(guideSet, guides = NULL, n_mismatches = 0,
  guide_length = 19, gc_content = c(0.4, 0.8), min_Son = 10,
  max_Soff = 50, consensus_range = NULL, n_clust = 11,
  method = "max", coeff = 1, PAM = "NGG", lower_count = 5,
  force = FALSE)
}
\arguments{
\item{guideSet}{guideSet object with targets.}

\item{guides}{Character. Optional vector of pre-computed guideRNA sequences to map and annotate. Will restrict downstream analysis to this set of guides.}

\item{n_mismatches}{Single integer of either 0, 1, 2, or 3. Maximal number of tolerated mismatches when assessing guideRNA binding targets. Defaults to 0.}

\item{guide_length}{Single integer between 15 and 25. Basepair size of the guideRNAs. Defaults to 19.}

\item{gc_content}{Numeric vector. Allowed GC content range of guides, e.g. c(0.4, 0.8) blacklists guides with GC content lower and higher than 40% and 80%, respectively}

\item{min_Son}{Numeric. Minimal on target score of guides. Defaults to 10.}

\item{consensus_range}{DataFrame with repname, start, and end columns. Discards guideRNAs targeting parts outside of \code{consensus_range} on the consensus.}

\item{n_clust}{Single positive integer <= 20. Number of groups to cluster guideRNAs into. 
Higher \code{n_clust} usually gives better results but comes with a speed penalty.}

\item{PAM}{Character. Currently only 'NGG' PAM is supported.}

\item{lower_count}{Numeric. Passed to jellyfish kmer counting. Only kmers occuring at least \code{lower_count} times are considered for mapping.}

\item{force}{Logical. If \code{TRUE}, overwrite existing guides.}

\item{min_Soff}{Numeric. Maximal off target score of guides. Defaults to 50.}
}
\value{
Returns a guideSet object containing guides.
}
\description{
Adds all potential guideRNAs against the targets of interest.
}
