% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTargets}
\alias{addTargets}
\title{Add targets to a guideSet object.}
\usage{
addTargets(guideSet, targets = NULL, blacklist = FALSE, min_dist = 0,
  force = FALSE)
}
\arguments{
\item{guideSet}{guideSet object containing genome annotation.}

\item{targets}{Character vector or GRanges object. Either identifier of repeat families or GRanges object with target coordinates.}

\item{blacklist}{Boolean. Should targets overlapping cis regulatory regions be blacklisted. False by default.}

\item{blacklist_dist}{Integer >= 0. Minimum basepair distance of target loci to cis regulatory regions.}

\item{force.}{If \code{TRUE}, overwrites existing results.}
}
\value{
Returns a guideSet object containing targets
}
\description{
Add targets to a guideSet object.
}
\examples{
guideSet <- addTargets(guideSet, targets = c('LTR12C', 'LTR12E'))
}
