% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTargets}
\alias{addTargets}
\title{Add targets to a guideSet object.}
\usage{
addTargets(guideSet, targets = NULL, force = FALSE)
}
\arguments{
\item{targets}{Character vector of family identifiers or GRanges object with coordinates and 'repname' metacolumn.}

\item{force}{Logical. If \code{TRUE}, overwrites existing results.}

\item{guideSet.}{guideSet object containing genome annotation.}
}
\value{
Returns a guideSet object containing targets.
}
\description{
Add targets to a guideSet object.
}
\examples{
\dontrun{
gs <- createGuideSet(Hsapiens, tes = te_anno)

# Use family identifiers as targets
gs <- addTargets(guideSet, targets = c('LTR12C', 'LTR12E'))

# Use custom regions as targets
target_coords <- GenomicRanges::makeGRangesFromDataFrame(data.frame('chrom'   = c('chr1', 'chr4'), 
                                                                    'start'   = c(1000, 10020),
                                                                    'end'     = c(1122, 10090),
                                                                    'strand'  = c('+', '-'),
                                                                    'repname' = c('ImaginativeTE', 'ImaginativeTE')),
                                                         , keep.extra.columns = TRUE)
gs <- addTargets(guideSet, targets = target_coords)
}
}
\seealso{
\code{\link[=plotTargets]{plotTargets()}}, and \code{\link[GenomicRanges:makeGRangesFromDataFrame]{GenomicRanges::makeGRangesFromDataFrame()}}
}
