% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guideR.R
\name{createGuideSet}
\alias{createGuideSet}
\title{Creates a new guideSet object}
\usage{
createGuideSet(genome, tes = GRanges(), cis = GRanges(), n_cores = 1,
  seed = 19)
}
\description{
Creates a new guideSet object
}
